+++ replacer-backend/replacer-domain/src/main/java/es/bvalero/replacer/page/save/WikipediaPageSaveRepository.java
@@ -6,5 +6,5 @@ import org.jmolecules.architecture.hexagonal.SecondaryPort;
 
 @SecondaryPort
 interface WikipediaPageSaveRepository {
-    void save(WikipediaPageSaveCommand pageSave, AccessToken accessToken) throws WikipediaException;
+    WikipediaPageSaveResult save(WikipediaPageSaveCommand pageSave, AccessToken accessToken) throws WikipediaException;
 }