+++ replacer-backend/replacer-web/src/test/java/es/bvalero/replacer/page/save/ReviewSaveControllerTest.java
@@ -96,6 +96,9 @@ class ReviewSaveControllerTest {
 
         when(applyCosmeticsService.applyCosmeticChanges(any(FinderPage.class))).thenReturn(pageSave.getContent());
         when(reviewSaveService.buildEditSummary(anyCollection(), anyBoolean())).thenReturn(pageSave.getEditSummary());
+        WikipediaTimestamp now = WikipediaTimestamp.now();
+        WikipediaPageSaveResult pageSaveResult = WikipediaPageSaveResult.builder().newTimestamp(now).build();
+        when(reviewSaveService.saveReviewContent(pageSave, user)).thenReturn(pageSaveResult);
 
         mvc
             .perform(
@@ -108,7 +111,7 @@ class ReviewSaveControllerTest {
             .andExpect(status().isNoContent());
 
         verify(reviewSaveService).saveReviewContent(pageSave, user);
-        verify(reviewSaveService).markAsReviewed(List.of(reviewedReplacement), true);
+        verify(reviewSaveService).markAsReviewed(List.of(reviewedReplacement), now);
     }
 
     @Test
@@ -127,7 +130,7 @@ class ReviewSaveControllerTest {
             .andExpect(status().isNoContent());
 
         verify(reviewSaveService, never()).saveReviewContent(any(WikipediaPageSaveCommand.class), any(User.class));
-        verify(reviewSaveService).markAsReviewed(List.of(reviewedReplacement), false);
+        verify(reviewSaveService).markAsReviewed(List.of(reviewedReplacement), null);
     }
 
     @Test
@@ -147,7 +150,7 @@ class ReviewSaveControllerTest {
             )
             .andExpect(status().isBadRequest());
 
-        verify(reviewSaveService, never()).markAsReviewed(anyCollection(), anyBoolean());
+        verify(reviewSaveService, never()).markAsReviewed(anyCollection(), any(WikipediaTimestamp.class));
     }
 
     @Test
@@ -171,7 +174,7 @@ class ReviewSaveControllerTest {
             .andExpect(status().isConflict());
 
         verify(reviewSaveService).saveReviewContent(pageSave, user);
-        verify(reviewSaveService, never()).markAsReviewed(anyCollection(), anyBoolean());
+        verify(reviewSaveService, never()).markAsReviewed(anyCollection(), any(WikipediaTimestamp.class));
     }
 
     @Test
@@ -197,7 +200,7 @@ class ReviewSaveControllerTest {
             .andExpect(status().isUnauthorized());
 
         verify(reviewSaveService).saveReviewContent(pageSave, user);
-        verify(reviewSaveService, never()).markAsReviewed(anyCollection(), anyBoolean());
+        verify(reviewSaveService, never()).markAsReviewed(anyCollection(), any(WikipediaTimestamp.class));
     }
 
     @Test
@@ -221,6 +224,6 @@ class ReviewSaveControllerTest {
             .andExpect(status().isInternalServerError());
 
         verify(reviewSaveService).saveReviewContent(pageSave, user);
-        verify(reviewSaveService, never()).markAsReviewed(anyCollection(), anyBoolean());
+        verify(reviewSaveService, never()).markAsReviewed(anyCollection(), any(WikipediaTimestamp.class));
     }
 }