+++ replacer-backend/replacer-wikipedia/src/main/java/es/bvalero/replacer/page/save/WikipediaPageSaveOfflineRepository.java
@@ -1,5 +1,6 @@
 package es.bvalero.replacer.page.save;
 
+import es.bvalero.replacer.page.find.WikipediaTimestamp;
 import es.bvalero.replacer.user.AccessToken;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Profile;
@@ -11,7 +12,16 @@ import org.springframework.stereotype.Service;
 class WikipediaPageSaveOfflineRepository implements WikipediaPageSaveRepository {
 
     @Override
-    public void save(WikipediaPageSaveCommand pageSave, AccessToken accessToken) {
-        // Do nothing
+    public WikipediaPageSaveResult save(WikipediaPageSaveCommand pageSave, AccessToken accessToken) {
+        return buildOfflineSaveResult();
+    }
+
+    private WikipediaPageSaveResult buildOfflineSaveResult() {
+        return WikipediaPageSaveResult
+            .builder()
+            .oldRevisionId(1)
+            .oldRevisionId(2)
+            .newTimestamp(WikipediaTimestamp.now())
+            .build();
     }
 }